requirements.txt
discord.py
openai
pandas
sqlalchemy